LayoutTests/fast/dom/DOMURL/url-very-large-password-no-crash.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
    }

    var url = new URL("about:blank");

    function runTest() {
        const largeSize = 800_000_000;
        const largeString = "\u0000".repeat(largeSize);

        description('Test setting the password attribute of a URL object with a very large value does not crash.');
        debug("Setting password");
        url.href = "http://@example.com/";
        url.password = largeString;
        shouldBe("url.href", "'http://example.com/'");
    }

    runTest();
    testRunner.notifyDone();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
